# algohive/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  competitions.CompetitionStatsResponse:
    properties:
      active_users:
        type: integer
      average_score:
        type: number
      competition_id:
        type: string
      completion_rate:
        type: number
      highest_score:
        type: number
      title:
        type: string
      total_users:
        type: integer
    type: object
...
  competitions.CreateTryRequest:
    properties:
      puzzle_id:
        type: string
      puzzle_index:
        type: integer
      puzzle_lvl:
        type: string
      step:
        type: integer
    required:
    - puzzle_id
    - puzzle_index
    - puzzle_lvl
    - step
    type: object
...
  competitions.UpdateTryRequest:
    properties:
      attempts:
        type: integer
      end_time:
        type: string
      score:
        type: number
    required:
    - attempts
    - end_time
    - score
    type: object
...
  models.Try:
    properties:
      attempts:
        type: integer
      competition_id:
        type: string
      end_time:
        type: string
      id:
        type: string
      puzzle_id:
        type: string
      puzzle_index:
        type: integer
      puzzle_lvl:
        type: string
      score:
        type: number
      start_time:
        type: string
      step:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact:
    email: ericphlpp@proton.me
    name: AlgoHive Support
  description: This is the API documentation for the AlgoHive API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger AlgoHive API
  version: 1.0.0
paths:
...
  /competitions/{id}/statistics:
    get:
      consumes:
      - application/json
      description: Get statistics for the specified competition
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/competitions.CompetitionStatsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get competition statistics
      tags:
      - Competitions
  /competitions/{id}/tries:
    get:
      consumes:
      - application/json
      description: Get all tries for the specified competition
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Try'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all tries for a competition
      tags:
      - Competitions
    post:
      consumes:
      - application/json
      description: Start a new try for a puzzle in a competition
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: string
      - description: Try details
        in: body
        name: try
        required: true
        schema:
          $ref: '#/definitions/competitions.CreateTryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Try'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Start a competition try
      tags:
      - Competitions
  /competitions/{id}/tries/{try_id}:
    put:
      consumes:
      - application/json
      description: Complete an ongoing try for a puzzle in a competition
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: string
      - description: Try ID
        in: path
        name: try_id
        required: true
        type: string
      - description: Try details
        in: body
        name: try
        required: true
        schema:
          $ref: '#/definitions/competitions.UpdateTryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Try'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Finish a competition try
      tags:
      - Competitions
  /competitions/{id}/users/{user_id}/tries:
    get:
      consumes:
      - application/json
      description: Get all tries for a specific user in the specified competition
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Try'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get user tries for a competition
      tags:
      - Competitions
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"